- removed a few warnings during VS2008 build

+++ trunk/src/hthread.cpp
@@ -127,7 +127,7 @@ namespace hltypes
 #endif
 	}
 	
-	void thread::sleep(int miliseconds)
+	void thread::sleep(float miliseconds)
 	{
 #ifdef _WIN32
 		Sleep(miliseconds);
